requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dnspython==2.3.0
docx==0.2.4
exceptiongroup==1.1.1
fastapi==0.97.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
loguru==0.7.0
lxml==4.9.2
multidict==6.0.4
numpy==1.19.0
openai==0.27.8
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
pinecone-client==2.2.1
pycparser==2.21
pydantic==1.10.9
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.1
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
Wand==0.6.11
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
